Reimplement binary_op

+++ src/main/java/org/fugalang/core/opcode/FrameOp.java
@@ -3,29 +3,29 @@ package org.fugalang.core.opcode;
 @SuppressWarnings("unused")
 public class FrameOp {
     // STACK OPERATIONS
-    public static final byte                 NOP =  1;
-    public static final byte             POP_TOP =  2;
-    public static final byte             ROT_TWO =  3;
-    public static final byte           ROT_THREE =  4;
-    public static final byte            ROT_FOUR =  5;
-    public static final byte             DUP_TOP =  6;
-    public static final byte         DUP_TOP_TWO =  7;
+    public static final int                 NOP =  1;
+    public static final int             POP_TOP =  2;
+    public static final int             ROT_TWO =  3;
+    public static final int           ROT_THREE =  4;
+    public static final int            ROT_FOUR =  5;
+    public static final int             DUP_TOP =  6;
+    public static final int         DUP_TOP_TWO =  7;
 
     // LOAD/STORE
 
-    public static final byte            LOAD_CONST =  8;
-    public static final byte            LOAD_NAME =  8;
-    public static final byte            LOAD_FAST =  8;
-    public static final byte            LOAD_ATTR =  8;
-    public static final byte            STORE_ITEM =  8;
-    public static final byte            STORE_ =  8;
-    public static final byte            UNARY_OP =  8;
-    public static final byte            RETURN_VALUE =  8;
+    public static final int            LOAD_CONST =  8;
+    public static final int            LOAD_NAME =  8;
+    public static final int            LOAD_FAST =  8;
+    public static final int            LOAD_ATTR =  8;
+    public static final int            STORE_ITEM =  8;
+    public static final int            STORE_ =  8;
+    public static final int            UNARY_OP =  8;
+    public static final int            RETURN_VALUE =  8;
 
-    public static final byte           UNARY_NOT =  9;
-    public static final byte           BINARY_OP = 11;
-    public static final byte          INPLACE_OP = 11;
-    public static final byte          PRINT_EXPR = 24;
+    public static final int           UNARY_NOT =  9;
+    public static final int           BINARY_OP = 11;
+    public static final int          INPLACE_OP = 11;
+    public static final int          PRINT_EXPR = 24;
 
     private FrameOp() {
     }
